more clean up for signer
